requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-core
faiss-cpu
pypdf
python-dotenv
streamlit
prometheus-client
flask